app/requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
bitsandbytes==0.41.2.post2
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
diskcache==5.6.3
dnspython==2.4.2
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
h11==0.14.0
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
kombu==5.3.4
llama_cpp_python==0.2.19
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai-whisper==20231117
packaging==23.2
peft==0.6.2
prompt-toolkit==3.0.41
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.12
websocket-client==1.6.4
yarl==1.9.3
